Rate Limiting added
Solving:
Brute Force Spam Attacks - Security Issue #6
Due to the lack of rate limiting, it is possible an attacker could spam the signup form to create thousands of accounts.

This has been fixed by implementing a session variable, updating if the user successfully creates an account, the cooldown time is currently set to 15 seconds, in a real-scenario this would be set much higher, however this is just for testing.