Brute Force Attacks are now prevented
Solving Issue "Brute Force Attacks still possible #12", where: Brute force password guessing is still possible for login fields, attackers are able to spam guess credentials of a user and gain unauthorised access to their account.

This is done through rate limiting, adding a maximum attempt limit of 5, preventing the user out from attempting to sign in again for 15 seconds if they fail to enter valid credentials 5 times. Cooldown is set to 15 seconds, in a real-scenario this would be set much higher, however this is just for testing. This has been successfully tested on the live server and works perfectly.