FAQ Chatbot
Implemented a simple FAQ Chatbot system that allows users to click a button at the bottom right of the page to open a chatbot interface.

This currently contains filler data, but will be updated with real FAQ questions a student / guest may have.

Chatbot is implemented onto every HTML page.