Automatic Table Creation
Solving GitHub issue "Automatic Table Creation #21", where:
As the database we are using is on MI-Linux and there is not 1 central database, but the database is localhosted and unique to the user, the tables have to be manually created to store information for forms such as signup and login. Instead of having to manually create these tables we can automate this process.

I have kept the initial database connection, as the database itself is always created, and added a query to check if the table to store user information already exists, and if not then this table is created.

Automatic table creation like this is fine for our project as we are working in a local development environment, however in industry, we would not be using localhosted databases but one central so this problem would be avoided. For our project, this simply saves time having to manually create the table on the database, providing better ease of access.